kdc: Make MAX_TIME INT32_MAX always
This just to be consistent with the new max value of HDB_entry's
max_life.  68 years is long enough.

+++ kdc/kerberos5.c
@@ -37,16 +37,16 @@
 #if SIZEOF_TIME_T == 4
 #define MAX_TIME ((time_t)INT32_MAX)
 #elif SIZEOF_TIME_T == 8
-#define MAX_TIME ((time_t)INT64_MAX)
+#define MAX_TIME ((time_t)INT32_MAX)
 #else
 #error "Unexpected sizeof(time_t)"
 #endif
 #else
 
 #if SIZEOF_TIME_T == 4
-#define MAX_TIME ((time_t)UINT32_MAX)
+#define MAX_TIME ((time_t)INT32_MAX)
 #else
-#define MAX_TIME ((time_t)UINT64_MAX)
+#define MAX_TIME ((time_t)INT32_MAX)
 #endif
 #endif
